+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 0.3.0 (2016-01-17)
+
+### Enhancements
+
+- Add image support in two positions (Left and Right).
+- Create enum `BHToastImagePosition`.
+- Add `messageAlignment` in `BHToastOptions`.
+- Change `messagePadding` to `contentInsets` in `BHToastOptions`.
+- Transfer `animationDuration` to `BHToastOptions`.
+
+
 ## 0.2.0 (2016-01-16)
 
 ### Enhancements